fixed gpu grid repeated interactions because of hash collisions

+++ src/tay/thread.c
@@ -101,14 +101,14 @@ void tay_runner_reset_stats() {
 }
 
 void tay_runner_report_stats() {
-    int broad_see_phase = 0;
-    int narrow_see_phase = 0;
+    unsigned long long broad_see_phase = 0;
+    unsigned long long narrow_see_phase = 0;
     for (int i = 0; i < runner.count; ++i) {
         TayThreadContext *c = &runner.threads[i].context;
         printf("  thread %d\n", i);
-        printf("    see phases: %d/%d\n", c->narrow_see_phase, c->broad_see_phase);
+        printf("    see phases: %llu/%llu\n", c->narrow_see_phase, c->broad_see_phase);
         broad_see_phase += c->broad_see_phase;
         narrow_see_phase += c->narrow_see_phase;
     }
-    printf("sum pf see phases: %d/%d\n", narrow_see_phase, broad_see_phase);
+    printf("sum pf see phases: %llu/%llu\n", narrow_see_phase, broad_see_phase);
 }
